refactor(timestamp): lazy-load the hover card and use IconButton for copy
Address review feedback:

- Restore the code split for the hover surface. HoverCard was imported
  statically, so every Timestamp bundled the overlay even with no card
  shown. Move the card (HoverCard + the copy rows) into a lazily-loaded
  TimestampHoverCard module and render it behind React.lazy + Suspense,
  only on the card-attached branch. The Suspense fallback is the bare
  <time>, so nothing disappears while the chunk loads. The copy row's
  Icon and IconButton ride the same split, so the default card-less
  Timestamp bundles no overlay or icon-button code.
- Replace the per-row hand-rolled <button> with the IconButton primitive
  (ghost, sm). Behavior is unchanged: click copies the row's value, the
  icon swaps copy -> check for the feedback window, a polite live region
  announces the copy, and the aria-label swaps between copy/copied. The
  timestamp-copy-button theme target now rides on the IconButton.
